requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
anthropic>=0.40.0
cryptography>=42.0.0
pydantic>=2.0.0
pytest>=8.0.0
